Home work 10

See merge request spinoza0/tfs-android-spring-2023!17

+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/data/network/model/message/MessageDto.kt
@@ -9,7 +9,7 @@ data class MessageDto(
     @SerialName("stream_id") val streamId: Long,
     @SerialName("sender_id") val senderId: Long,
     @SerialName("content") val content: String,
-    @SerialName("recipient_id") val recipientId: Int,
+    @SerialName("recipient_id") val recipientId: Long,
     @SerialName("timestamp") val timestamp: Long,
     @SerialName("subject") val subject: String,
     @SerialName("is_me_message") val isMeMessage: Boolean,
